feat(system2d): remove system2d.py
Removed old solver functions.
